refactor: QrProperties 패키지 경로 변경 및 retryable 제거
