+++ backend/build.gradle.kts
@@ -29,7 +29,6 @@ dependencies {
     implementation("org.springframework.boot:spring-boot-starter-websocket")
     implementation("org.springframework.boot:spring-boot-starter-aop")
     implementation("org.springframework.boot:spring-boot-starter-actuator")
-    implementation("org.springframework.retry:spring-retry")
     implementation("io.github.20hyeonsulee:websocket-docs-generator:1.0.7")
 
     implementation("com.google.zxing:core:3.5.3")
