+++ backend/src/test/java/coffeeshout/global/config/IntegrationTestConfig.java
@@ -2,7 +2,6 @@ package coffeeshout.global.config;
 
 import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Import;
 import org.springframework.context.annotation.Profile;
 import org.springframework.scheduling.TaskScheduler;
